requirements.txt
aniso8601==8.1.0
attrs==20.3.0
blinker==1.4
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT-Extended==3.24.1
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.4
google-api-core==1.24.1
google-auth==1.24.0
google-cloud-core==1.5.0
google-cloud-logging==1.15.1
google-cloud-secret-manager==2.4.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.4.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
nltk==3.6.2
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
PyMySQL==0.10.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.25.1
rsa==4.7
sentry-sdk==1.1.0
six==1.15.0
SQLAlchemy==1.3.18
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==0.16.1
wordsegment==1.3.1
zipp==3.4.0
